+++ appyx-components/experimental/cards/common/src/commonMain/kotlin/com/bumble/appyx/components/experimental/cards/operation/VotePass.kt
@@ -2,10 +2,10 @@ package com.bumble.appyx.components.experimental.cards.operation
 
 import androidx.compose.animation.core.AnimationSpec
 import com.bumble.appyx.components.experimental.cards.Cards
-import com.bumble.appyx.interactions.Parcelize
-import com.bumble.appyx.interactions.core.model.transition.Operation
 import com.bumble.appyx.components.experimental.cards.CardsModel
 import com.bumble.appyx.components.experimental.cards.CardsModel.State.Card.InvisibleCard.VotedCard.VOTED_CARD_STATE.PASSED
+import com.bumble.appyx.interactions.core.model.transition.Operation
+import com.bumble.appyx.utils.multiplatform.Parcelize
 
 @Parcelize
 class VotePass<InteractionTarget>(