+++ appyx-components/experimental/modal/common/src/commonMain/kotlin/com/bumble/appyx/components/modal/ModalModel.kt
@@ -1,12 +1,12 @@
 package com.bumble.appyx.components.modal
 
-import com.bumble.appyx.interactions.Parcelable
-import com.bumble.appyx.interactions.Parcelize
 import com.bumble.appyx.interactions.core.Element
 import com.bumble.appyx.interactions.core.Elements
 import com.bumble.appyx.interactions.core.asElement
 import com.bumble.appyx.interactions.core.model.transition.BaseTransitionModel
 import com.bumble.appyx.interactions.core.state.SavedStateMap
+import com.bumble.appyx.utils.multiplatform.Parcelable
+import com.bumble.appyx.utils.multiplatform.Parcelize
 
 class ModalModel<InteractionTarget : Any>(
     initialElements: List<InteractionTarget>,