+++ appyx-interactions/common/src/commonMain/kotlin/com/bumble/appyx/interactions/core/model/transition/BaseTransitionModel.kt
@@ -1,13 +1,15 @@
 package com.bumble.appyx.interactions.core.model.transition
 
-import com.bumble.appyx.interactions.AppyxLogger
-import com.bumble.appyx.interactions.Parcelable
 import com.bumble.appyx.interactions.core.Element
-import com.bumble.appyx.interactions.core.model.transition.Operation.Mode.*
+import com.bumble.appyx.interactions.core.model.transition.Operation.Mode.IMMEDIATE
+import com.bumble.appyx.interactions.core.model.transition.Operation.Mode.IMPOSED
+import com.bumble.appyx.interactions.core.model.transition.Operation.Mode.KEYFRAME
 import com.bumble.appyx.interactions.core.model.transition.TransitionModel.Output
 import com.bumble.appyx.interactions.core.model.transition.TransitionModel.SettleDirection
 import com.bumble.appyx.interactions.core.state.MutableSavedStateMap
 import com.bumble.appyx.interactions.core.state.SavedStateMap
+import com.bumble.appyx.utils.multiplatform.AppyxLogger
+import com.bumble.appyx.utils.multiplatform.Parcelable
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -117,7 +119,7 @@ abstract class BaseTransitionModel<InteractionTarget, ModelState : Parcelable>(
         val baseLine = state.value
 
         return if (operation.isApplicable(baseLine.currentTargetState)) {
-            val transition = operation.invoke(baseLine.currentTargetState.removeDestroyedElements())
+            val transition = operation.invoke(baseLine.currentTargetState)
             val newState = baseLine.deriveUpdate(transition)
             updateState(newState)
             true