+++ appyx-interactions/common/src/commonMain/kotlin/com/bumble/appyx/interactions/platform.kt
@@ -1,13 +1,3 @@
 package com.bumble.appyx.interactions
 
 expect fun getPlatformName(): String
-
-@Target(AnnotationTarget.CLASS)
-@Retention(AnnotationRetention.BINARY)
-expect annotation class Parcelize()
-
-@Target(AnnotationTarget.TYPE)
-@Retention(AnnotationRetention.BINARY)
-expect annotation class RawValue()
-
-expect interface Parcelable