+++ appyx-navigation/common/src/commonMain/kotlin/com/bumble/appyx/navigation/platform/PlatformLifecycleRegistry.kt
@@ -1,0 +1,12 @@
+package com.bumble.appyx.navigation.platform
+
+import com.bumble.appyx.navigation.lifecycle.CommonLifecycleOwner
+import com.bumble.appyx.navigation.lifecycle.Lifecycle
+
+expect class PlatformLifecycleRegistry : Lifecycle {
+    fun setCurrentState(state: Lifecycle.State)
+
+    companion object {
+        fun create(owner: CommonLifecycleOwner): PlatformLifecycleRegistry
+    }
+}