+++ demos/appyx-interactions/desktop/src/desktopMain/kotlin/com/bumble/appyx/interactions/widgets/Widgets.kt
@@ -110,7 +110,7 @@ private fun WidgetsUi(
                     .fillMaxWidth()
                     .requiredHeight(240.dp)
             )
-        }
+        },
     )
 }