+++ demos/mkdocs/appyx-interactions/interactions/sample2/web/src/jsMain/kotlin/com/bumble/appyx/demos/sample2/Sample2MotionController.kt
@@ -1,7 +1,6 @@
 package com.bumble.appyx.demos.sample2
 
 import androidx.compose.animation.core.SpringSpec
-import com.bumble.appyx.interactions.core.ui.property.impl.Position.Alignment
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.unit.Density
 import androidx.compose.ui.unit.DpOffset
@@ -12,7 +11,6 @@ import com.bumble.appyx.components.internal.testdrive.TestDriveModel.State.Eleme
 import com.bumble.appyx.components.internal.testdrive.TestDriveModel.State.ElementState.C
 import com.bumble.appyx.components.internal.testdrive.TestDriveModel.State.ElementState.D
 import com.bumble.appyx.components.internal.testdrive.operation.MoveTo
-import com.bumble.appyx.interactions.AppyxLogger
 import com.bumble.appyx.interactions.core.ui.context.TransitionBounds
 import com.bumble.appyx.interactions.core.ui.context.UiContext
 import com.bumble.appyx.interactions.core.ui.gesture.Drag.Direction8.DOWN
@@ -29,9 +27,11 @@ import com.bumble.appyx.interactions.core.ui.gesture.dragDirection8
 import com.bumble.appyx.interactions.core.ui.helper.DefaultAnimationSpec
 import com.bumble.appyx.interactions.core.ui.property.impl.BackgroundColor
 import com.bumble.appyx.interactions.core.ui.property.impl.Position
+import com.bumble.appyx.interactions.core.ui.property.impl.Position.Alignment
 import com.bumble.appyx.interactions.core.ui.property.impl.RotationZ
 import com.bumble.appyx.interactions.core.ui.state.MatchedTargetUiState
 import com.bumble.appyx.transitionmodel.BaseMotionController
+import com.bumble.appyx.utils.multiplatform.AppyxLogger
 
 class Sample2MotionController<InteractionTarget : Any>(
     uiContext: UiContext,
@@ -83,7 +83,10 @@ class Sample2MotionController<InteractionTarget : Any>(
         )
     }
 
-    override fun mutableUiStateFor(uiContext: UiContext, targetUiState: TargetUiState): MutableUiState =
+    override fun mutableUiStateFor(
+        uiContext: UiContext,
+        targetUiState: TargetUiState
+    ): MutableUiState =
         targetUiState.toMutableState(uiContext)
 
 
@@ -98,10 +101,10 @@ class Sample2MotionController<InteractionTarget : Any>(
         ): Gesture<InteractionTarget, TestDriveModel.State<InteractionTarget>> {
             // FIXME 60.dp is the assumed element size, connect it to real value
             // TODO automate this whole calculation based on .onPlaced centers of targetUiStates
-            val maxX = with (density) {
+            val maxX = with(density) {
                 (transitionBounds.widthDp - 60.dp).toPx()
             }
-            val maxY = with (density) {
+            val maxY = with(density) {
                 (transitionBounds.heightDp + bottomOffset.y - 60.dp).toPx()
             }