+++ demos/mkdocs/appyx-interactions/interactions/sample3/web/src/jsMain/kotlin/com/bumble/appyx/demos/sample3/Sample3MotionController.kt
@@ -1,8 +1,6 @@
 package com.bumble.appyx.demos.sample3
 
-import com.bumble.appyx.interactions.core.ui.helper.DefaultAnimationSpec
 import androidx.compose.animation.core.SpringSpec
-import com.bumble.appyx.interactions.core.ui.property.impl.Position.Alignment
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.unit.Density
 import androidx.compose.ui.unit.DpOffset
@@ -13,7 +11,6 @@ import com.bumble.appyx.components.internal.testdrive.TestDriveModel.State.Eleme
 import com.bumble.appyx.components.internal.testdrive.TestDriveModel.State.ElementState.C
 import com.bumble.appyx.components.internal.testdrive.TestDriveModel.State.ElementState.D
 import com.bumble.appyx.components.internal.testdrive.operation.MoveTo
-import com.bumble.appyx.interactions.AppyxLogger
 import com.bumble.appyx.interactions.core.ui.context.TransitionBounds
 import com.bumble.appyx.interactions.core.ui.context.UiContext
 import com.bumble.appyx.interactions.core.ui.gesture.Drag.Direction8.DOWN
@@ -27,10 +24,13 @@ import com.bumble.appyx.interactions.core.ui.gesture.Drag.Direction8.UPRIGHT
 import com.bumble.appyx.interactions.core.ui.gesture.Gesture
 import com.bumble.appyx.interactions.core.ui.gesture.GestureFactory
 import com.bumble.appyx.interactions.core.ui.gesture.dragDirection8
+import com.bumble.appyx.interactions.core.ui.helper.DefaultAnimationSpec
 import com.bumble.appyx.interactions.core.ui.property.impl.BackgroundColor
 import com.bumble.appyx.interactions.core.ui.property.impl.Position
+import com.bumble.appyx.interactions.core.ui.property.impl.Position.Alignment
 import com.bumble.appyx.interactions.core.ui.state.MatchedTargetUiState
 import com.bumble.appyx.transitionmodel.BaseMotionController
+import com.bumble.appyx.utils.multiplatform.AppyxLogger
 
 class Sample3MotionController<InteractionTarget : Any>(
     uiContext: UiContext,
@@ -78,7 +78,10 @@ class Sample3MotionController<InteractionTarget : Any>(
         )
     }
 
-    override fun mutableUiStateFor(uiContext: UiContext, targetUiState: TargetUiState): MutableUiState =
+    override fun mutableUiStateFor(
+        uiContext: UiContext,
+        targetUiState: TargetUiState
+    ): MutableUiState =
         targetUiState.toMutableState(uiContext)
 
     class Gestures<InteractionTarget>(
@@ -92,10 +95,10 @@ class Sample3MotionController<InteractionTarget : Any>(
         ): Gesture<InteractionTarget, TestDriveModel.State<InteractionTarget>> {
             // FIXME 60.dp is the assumed element size, connect it to real value
             // TODO automate this whole calculation based on .onPlaced centers of targetUiStates
-            val maxX = with (density) {
+            val maxX = with(density) {
                 (transitionBounds.widthDp - 60.dp).toPx()
             }
-            val maxY = with (density) {
+            val maxY = with(density) {
                 (transitionBounds.heightDp + bottomOffset.y - 60.dp).toPx()
             }
 
@@ -107,18 +110,21 @@ class Sample3MotionController<InteractionTarget : Any>(
                     DOWN -> Gesture(MoveTo(D), Offset(0f, maxY))
                     else -> Gesture.Noop()
                 }
+
                 B -> when (direction) {
                     DOWN -> Gesture(MoveTo(C), Offset(0f, maxY))
                     DOWNLEFT -> Gesture(MoveTo(D), Offset(-maxX, maxY))
                     LEFT -> Gesture(MoveTo(A), Offset(-maxX, 0f))
                     else -> Gesture.Noop()
                 }
+
                 C -> when (direction) {
                     LEFT -> Gesture(MoveTo(D), Offset(-maxX, 0f))
                     UPLEFT -> Gesture(MoveTo(A), Offset(-maxX, -maxY))
                     UP -> Gesture(MoveTo(B), Offset(0f, -maxY))
                     else -> Gesture.Noop()
                 }
+
                 D -> when (direction) {
                     UP -> Gesture(MoveTo(A), Offset(0f, -maxY))
                     UPRIGHT -> Gesture(MoveTo(B), Offset(maxX, -maxY))