+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Pending changes
 
+- [#551](https://github.com/bumble-tech/appyx/pull/551) - Multiplatform Navigation module
+
+### Fixed
+
+- [#560](https://github.com/bumble-tech/appyx/issues/560) – Don't remove destroyed elements on update automatically
+
+---
+
+## 2.0.0-alpha02
+
 ### Added
 
 - [#539](https://github.com/bumble-tech/appyx/pull/539) – Position alignment
@@ -10,6 +20,8 @@
 ### Fixed
 
 - [#530](https://github.com/bumble-tech/appyx/issues/530) – Backstack Parallax motion controller bug
+
+<div style="text-align: center"><small>9 Aug 2023</small></div>
 
 ---